+++ api/include/opentelemetry/plugin/detail/loader_info.h
