+++ api/include/opentelemetry/plugin/hook.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "opentelemetry/plugin/factory.h"
-#include "opentelemetry/plugin/loader_info.h"
+#include "opentelemetry/plugin/detail/loader_info.h"
 
 #ifdef WIN32
